requirements.txt
apache-flink==1.17.1
click
psycopg2-binary